+++ lib/datamappify/data.rb
@@ -1,6 +1,4 @@
 require 'datamappify/data/errors'
-require 'datamappify/data/util'
-require 'datamappify/data/persistence'
 
 module Datamappify
   module Data