+++ lib/datamappify/data/provider/active_record/base.rb
@@ -1,9 +1,0 @@
-module Datamappify
-  module Data
-    module ActiveRecord
-      class Base < ::ActiveRecord::Base
-        self.abstract_class = true
-      end
-    end
-  end
-end