+++ spec/repository/persistence_spec.rb
@@ -16,9 +16,11 @@ shared_examples_for "repository persistence" do |data_provider|
 
   describe "#find" do
     describe "resource" do
+      let!(:user) { user_repository.save(existing_user) }
+
       it "found" do
-        user_repository.find(1).should == existing_user
-        user_repository.find([1]).should == [existing_user]
+        user_repository.find(user.id).should == existing_user
+        user_repository.find([user.id]).should == [existing_user]
       end
 
       it "not found" do
@@ -27,12 +29,14 @@ shared_examples_for "repository persistence" do |data_provider|
     end
 
     describe "collection" do
+      let!(:user) { user_repository.save(existing_user) }
+
       it "found" do
-        user_repository.find([1]).should == [existing_user]
+        user_repository.find([user.id]).should == [existing_user]
       end
 
       it "partial found" do
-        user_repository.find([1, 2]).should == [existing_user]
+        user_repository.find([user.id, 42]).should == [existing_user]
       end
 
       it "not found" do
@@ -43,9 +47,10 @@ shared_examples_for "repository persistence" do |data_provider|
 
   describe "#save" do
     it "success" do
-      expect { user_repository.save(new_valid_user) }.to change { user_repository.count }.by(1)
+      new_user = nil
 
-      new_user = user_repository.find(user_repository.count)
+      expect { new_user = user_repository.save(new_valid_user) }.to change { user_repository.count }.by(1)
+
       new_user.should be_kind_of(User)
       new_user.first_name.should == 'Batman'
       new_user.driver_license.should == 'ARKHAMCITY'
@@ -56,15 +61,13 @@ shared_examples_for "repository persistence" do |data_provider|
     end
 
     it "transaction" do
-      DatabaseCleaner.clean
-
       new_valid_user.passport = 'DEVOPS'
 
-      -> { user_repository.save(new_valid_user) }.should raise_error(ActiveRecord::StatementInvalid)
+      save_action = Proc.new { -> { user_repository.save(new_valid_user) }.should raise_error }
 
-      user_repository.count.should == 0
-      data_passports.count.should == 0
-      data_driver_licenses.count.should == 0
+      expect { save_action.call }.to change { user_repository.count }.by(0)
+      expect { save_action.call }.to change { data_passports.count }.by(0)
+      expect { save_action.call }.to change { data_driver_licenses.count }.by(0)
     end
 
     describe "update multiple entities" do
@@ -104,7 +107,9 @@ shared_examples_for "repository persistence" do |data_provider|
         user.first_name = 'Vivian'
         user.driver_license = 'LOCOMOTE'
 
-        updated_user = user_repository.save(user)
+        updated_user = nil
+
+        expect { updated_user = user_repository.save(user) }.to change { user_repository.count }.by(0)
 
         updated_user.first_name.should == 'Vivian'
         updated_user.driver_license.should == 'LOCOMOTE'
@@ -113,8 +118,6 @@ shared_examples_for "repository persistence" do |data_provider|
 
         persisted_user.first_name.should == 'Vivian'
         persisted_user.driver_license.should == 'LOCOMOTE'
-
-        user_repository.count.should == 1
       end
 
       it "updates existing and new records" do
@@ -123,7 +126,9 @@ shared_examples_for "repository persistence" do |data_provider|
         user.first_name = 'Vivian'
         user.health_care = 'BATMANCAVE'
 
-        updated_user = user_repository.save(user)
+        updated_user = nil
+
+        expect { updated_user = user_repository.save(user) }.to change { user_repository.count }.by(0)
 
         updated_user.first_name.should == 'Vivian'
         updated_user.health_care.should == 'BATMANCAVE'
@@ -132,42 +137,32 @@ shared_examples_for "repository persistence" do |data_provider|
 
         persisted_user.first_name.should == 'Vivian'
         persisted_user.health_care.should == 'BATMANCAVE'
-
-        user_repository.count.should == 1
       end
     end
   end
 
   describe "#destroy" do
     describe "resource" do
+      let!(:user) { user_repository.save(new_valid_user) }
+
       it "via id" do
-        user_repository.destroy!(1)
-        user_repository.count.should == 0
+        expect { user_repository.destroy!(user.id) }.to change { user_repository.count }.by(-1)
       end
 
       it "via entity" do
-        user = user_repository.find(1)
-
-        user_repository.destroy!(user)
-        user_repository.count.should == 0
+        expect { user_repository.destroy!(user) }.to change { user_repository.count }.by(-1)
       end
     end
 
     describe "collection" do
-      before do
-        user_repository.save(new_valid_user)
-      end
+      let!(:users) { user_repository.save([new_valid_user, new_valid_user2]) }
 
       it "via id" do
-        user_repository.destroy!([1, 2])
-        user_repository.count.should == 0
+        expect { user_repository.destroy!(users.map(&:id)) }.to change { user_repository.count }.by(-2)
       end
 
       it "via entity" do
-        users = user_repository.find([1, 2])
-
-        user_repository.destroy!(users)
-        user_repository.count.should == 0
+        expect { user_repository.destroy!(users) }.to change { user_repository.count }.by(-2)
       end
     end
   end
@@ -176,5 +171,25 @@ end
 describe Datamappify::Repository do
   DATA_PROVIDERS.each do |data_provider|
     it_behaves_like "repository persistence", data_provider
+  end
+
+  describe "entity composed from multiple data providers" do
+    let!(:hero_repository) { HeroUserRepository.instance }
+    let!(:hero)            { HeroUser.new :first_name => 'Aaron', :last_name => 'Patterson' }
+
+    it "#find" do
+      saved_hero = nil
+
+      expect { saved_hero = hero_repository.save!(hero) }.to change { hero_repository.count }.by(1)
+
+      saved_hero.first_name.should == 'Aaron'
+      saved_hero.last_name.should == 'Patterson'
+      saved_hero.nickname = 'unsaved nickname'
+
+      persisted_hero = hero_repository.find(saved_hero.id)
+      persisted_hero.first_name.should == 'Aaron'
+      persisted_hero.last_name.should == 'Patterson'
+      persisted_hero.nickname.should be_nil
+    end
   end
 end