+++ spec/support/repositories/sequel/comment_repository.rb
@@ -1,0 +1,6 @@
+class CommentRepositorySequel
+  include Datamappify::Repository
+
+  for_entity Comment
+  default_provider :Sequel
+end