+++ spec/support/repositories/sequel/group_repository.rb
@@ -1,0 +1,6 @@
+class GroupRepositorySequel
+  include Datamappify::Repository
+
+  for_entity Group
+  default_provider :Sequel
+end