+++ spec/support/repositories/sequel/role_repository.rb
@@ -1,0 +1,6 @@
+class RoleRepositorySequel
+  include Datamappify::Repository
+
+  for_entity Role
+  default_provider :Sequel
+end