+++ spec/support/tables/active_record.rb
@@ -5,11 +5,10 @@ ActiveRecord::Migration.suppress_messages do
   )
 
   ActiveRecord::Schema.define(:version => 0) do
-    begin
-      drop_table :users
-      drop_table :comments
-      drop_table :roles
-    rescue
+    create_table :hero_users do |t|
+      t.string :first_name, :null => false
+      t.string :nickname
+      t.timestamps
     end
 
     create_table :users do |t|